requirements.txt
runpod
pydantic>=2
pillow
opencv-python-headless
numpy
